pyproject.toml
[tool.poetry]
name = "es-site2"
version = "0.1.0"
description = ""
authors = ["STI-ESUFRN"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.6"
djangorestframework = "^3.13.1"
django-multiselectfield = "^0.1.12"
gunicorn = "^20.1.0"
django-ckeditor = "^6.5.0"
mysqlclient = "^2.1.1"
python-decouple = "^3.6"
django-htmlmin = "^0.11.0"
django-admin-interface = "^0.20.0"
python-dateutil = "^2.8.2"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
